feat: 시작 화면에서 Exam, Quiz
